move func `do_shotnum_intersection` from module hdfreadcontrol.py to helpers.py (update test routines accordingly
