+++ bapsflib/_hdf/utils/helpers.py
@@ -23,6 +23,7 @@ from .file import File
 
 # define type aliases
 ControlMap = Union[HDFMapControlTemplate, HDFMapControlCLTemplate]
+IndexDict = Dict[str, np.ndarray]
 
 
 def build_shotnum_dset_relation(
@@ -579,3 +580,47 @@ def condition_shotnum(shotnum: Any,
 
     # return
     return shotnum
+
+
+def do_shotnum_intersection(
+        shotnum: np.ndarray,
+        sni_dict: IndexDict,
+        index_dict: IndexDict) -> Tuple[np.ndarray, IndexDict,
+                                        IndexDict]:
+    """
+    Calculates intersection of **shotnum** and all existing dataset
+    shot numbers, **shotnum[sni]**.
+
+    :param shotnum: desired HDF5 shot numbers
+    :param sni_dict: dictionary of all dataset **sni** arrays
+    :param index_dict:  dictionary of all dataset **index** arrays
+    :return: intersected and re-calculated versions of :code:`index`
+        and :code:`sni` numpy arrays
+
+    .. admonition:: Recall Array Relationship
+
+        .. code-block:: python
+
+            shotnum[sni] = dset[index, shotnumkey]
+    """
+    # intersect shot numbers
+    shotnum_intersect = shotnum
+    for sni in sni_dict.values():
+        shotnum_intersect = np.intersect1d(shotnum_intersect,
+                                           shotnum[sni],
+                                           assume_unique=True)
+    if shotnum_intersect.shape[0] == 0:
+        raise ValueError('Input `shotnum` would result in a NULL array')
+
+    # now filter
+    for cname in index_dict:
+        sni = sni_dict[cname]
+        mask_for_index = np.isin(shotnum[sni], shotnum_intersect)
+        index_dict[cname] = index_dict[cname][mask_for_index]
+        sni_dict[cname] = np.ones(shotnum_intersect.shape, dtype=bool)
+
+    # update shotnum
+    shotnum = shotnum_intersect
+
+    # return
+    return shotnum, sni_dict, index_dict
