+++ bapsflib/_hdf/utils/tests/test_hdfreadcontrol.py
@@ -18,275 +18,12 @@ import unittest as ut
 
 from bapsflib._hdf.maps import (ConType, HDFMap)
 from bapsflib._hdf.maps.controls.templates import HDFMapControlTemplate
-from bapsflib._hdf.maps.controls.waveform import HDFMapControlWaveform
-from numpy.lib import recfunctions as rfn
 from typing import (Any, Dict, List, Tuple)
 from unittest import mock
 
 from . import TestBase
 from ..file import File
-from ..hdfreadcontrol import (build_shotnum_dset_relation,
-                              do_shotnum_intersection,
-                              HDFReadControl)
-
-
-class TestBuildShotnumDsetRelation(TestBase):
-    """Test Case for build_shotnum_dset_relation"""
-
-    def setUp(self):
-        # setup HDF5 file
-        super().setUp()
-        self.f.add_module('Waveform',
-                          mod_args={'n_configs': 1, 'sn_size': 100})
-        self.mod = self.f.modules['Waveform']
-
-    def tearDown(self):
-        super().tearDown()
-
-    @property
-    def cgroup(self):
-        return self.f['Raw data + config/Waveform']
-
-    @property
-    def map(self):
-        return HDFMapControlWaveform(self.cgroup)
-
-    def test_simple_dataset(self):
-        """
-        Tests for a dataset containing recorded data for a single
-        configuration.
-        """
-        # -- dset with 1 shotnum                                    ----
-        self.mod.knobs.sn_size = 1
-        self.assertInRangeSN()
-        self.assertOutRangeSN()
-
-        # -- typical dset with sequential shot numbers              ----
-        self.mod.knobs.sn_size = 100
-        self.assertInRangeSN()
-        self.assertOutRangeSN()
-
-        # -- dset with non-sequential shot numbers                  ----
-        self.mod.knobs.sn_size = 100
-        data = self.cgroup['Run time list'][...]
-        data['Shot number'] = np.append(
-            np.arange(5, 25, dtype=np.uint32),
-            np.append(np.arange(51, 111, dtype=np.uint32),
-                      np.arange(150, 170, dtype=np.uint32)))
-        del self.cgroup['Run time list']
-        self.cgroup.create_dataset('Run time list', data=data)
-        self.assertInRangeSN()
-        self.assertOutRangeSN()
-
-    def test_complex_dataset(self):
-        """
-        Tests for a dataset containing recorded data for a multiple
-        configurations.
-        """
-        # define multiple configurations for one dataset
-        self.mod.knobs.n_configs = 3
-
-        # -- dset with 1 shotnum                                    ----
-        self.mod.knobs.sn_size = 1
-        self.assertInRangeSN()
-        self.assertOutRangeSN()
-
-        # -- typical dset with sequential shot numbers              ----
-        self.mod.knobs.sn_size = 100
-        self.assertInRangeSN()
-        self.assertOutRangeSN()
-
-        # -- dset with non-sequential shot numbers                  ----
-        self.mod.knobs.sn_size = 100
-        data = self.cgroup['Run time list'][...]
-        sn_arr = np.append(
-            np.arange(5, 25, dtype=np.uint32),
-            np.append(np.arange(51, 111, dtype=np.uint32),
-                      np.arange(150, 170, dtype=np.uint32)))
-        data['Shot number'][0::3] = sn_arr
-        data['Shot number'][1::3] = sn_arr
-        data['Shot number'][2::3] = sn_arr
-        del self.cgroup['Run time list']
-        self.cgroup.create_dataset('Run time list', data=data)
-        self.assertInRangeSN()
-        self.assertOutRangeSN()
-
-        # -- dset without a configuration fields                    ----
-        self.mod.knobs.sn_size = 50
-        data = self.cgroup['Run time list'][...]
-        data = rfn.rename_fields(data, {'Configuration name': 'oops'})
-        del self.cgroup['Run time list']
-        self.cgroup.create_dataset('Run time list', data=data)
-        cdset = self.cgroup['Run time list']
-        with self.assertRaises(ValueError):
-            build_shotnum_dset_relation(
-                np.empty(5, dtype=np.uint32), cdset, 'Shot number',
-                self.map, 'config01')
-
-    def assertInRangeSN(self):
-        """
-        Assert shot numbers cases with in-range of dataset shot numbers.
-        """
-        shotnum_list = [
-            [10],
-            [50, 51],
-            [50, 60],
-            [1, self.mod.knobs.sn_size],
-            [50, 51, 52, 53, 54, 55, 56, 57, 58, 59],
-            [1, 11, 21, 31, 41, 51, 61, 71, 81, 91]
-        ]
-        cdset = self.cgroup['Run time list']
-        shotnumkey = 'Shot number'
-        configkey = 'Configuration name'
-        for og_shotnum in shotnum_list:
-            if og_shotnum == [1, 1]:
-                continue
-
-            sn_arr = np.array(og_shotnum, dtype=np.uint32)
-            for cconfn in self.map.configs:
-                index, sni = build_shotnum_dset_relation(
-                    sn_arr, cdset, shotnumkey, self.map, cconfn)
-
-                self.assertSNSuite(sn_arr, index, sni,
-                                   cdset, shotnumkey,
-                                   configkey, cconfn)
-
-    def assertOutRangeSN(self):
-        """
-        Assert shot number cases where some shot numbers are out of
-        range of the dataset shotnumbers.
-        """
-        # Note: condition_shotnum() will ensure shotnum >= 0 so
-        #       build_shotnum_dset_relation() does not handle this
-        #
-        # - one above largest shot number
-        # - out of range above (sn_size+1, sn_size+10, sn_size+100)
-        #   and valid
-        #
-        shotnum_list = [
-            [self.mod.knobs.sn_size + 1],
-            [self.mod.knobs.sn_size + 1, self.mod.knobs.sn_size + 10],
-            [10, 15, self.mod.knobs.sn_size + 1],
-            [10, 15, self.mod.knobs.sn_size + 1,
-             self.mod.knobs.sn_size + 10,
-             self.mod.knobs.sn_size + 100],
-        ]
-        cdset = self.cgroup['Run time list']
-        shotnumkey = 'Shot number'
-        configkey = 'Configuration name'
-        for og_shotnum in shotnum_list:
-            sn_arr = np.array(og_shotnum, dtype=np.uint32)
-            for cconfn in self.map.configs:
-                index, sni = build_shotnum_dset_relation(
-                    sn_arr, cdset, shotnumkey, self.map, cconfn)
-
-                self.assertSNSuite(sn_arr, index, sni,
-                                   cdset, shotnumkey,
-                                   configkey, cconfn)
-
-    def assertSNSuite(self, shotnum, index, sni,
-                      cdset, shotnumkey, configkey, cconfn):
-        """Suite of assertions for shot number conditioning"""
-        # shotnum    - original requested shot number
-        # index      - index of dataset
-        # sni        - boolean mask for shotnum
-        #               shotnum[sni] = cdset[index, shotnumkey]
-        # cdset      - control devices dataset
-        # shotnumkey - field in cdset that corresponds to shot numbers
-        # configkey  - field in cdset that corresponds to configuration
-        #              names
-        # cconfn     - configuration name for control device
-        #
-        # all return variables should be np.ndarray
-        self.assertTrue(isinstance(index, np.ndarray))
-        self.assertTrue(isinstance(sni, np.ndarray))
-
-        # all should be 1D arrays
-        self.assertEqual(index.ndim, 1)
-        self.assertEqual(sni.ndim, 1)
-
-        # equate array sizes
-        self.assertEqual(shotnum.size, sni.size)
-        self.assertEqual(np.count_nonzero(sni), index.size)
-
-        # shotnum[sni] = cdset[index, shotnumkey]
-        if index.size != 0:
-            self.assertTrue(np.array_equal(
-                shotnum[sni], cdset[index.tolist(), shotnumkey]))
-        else:
-            self.assertEqual(shotnum[sni].size, 0)
-
-        # ensure correct config is grabbed
-        if index.size != 0:
-            cname_arr = cdset[index.tolist(), configkey]
-            for name in cname_arr:
-                self.assertEqual(name.decode('utf-8'), cconfn)
-
-
-class TestDoShotnumIntersection(ut.TestCase):
-    """Test Case for do_shotnum_intersection"""
-    def test_one_control(self):
-        """Test intersection behavior with one control device"""
-        # test a case that results in a null result
-        shotnum = np.arange(1, 21, 1, dtype=np.uint32)
-        sni_dict = {'Waveform': np.zeros(shotnum.shape, dtype=bool)}
-        index_dict = {'Waveform': np.array([])}
-        self.assertRaises(ValueError,
-                          do_shotnum_intersection,
-                          shotnum, sni_dict, index_dict)
-
-        # test a working case
-        shotnum = np.arange(1, 21, 1)
-        sni_dict = {'Waveform': np.zeros(shotnum.shape, dtype=bool)}
-        index_dict = {'Waveform': np.array([5, 6, 7])}
-        sni_dict['Waveform'][[5, 6, 7]] = True
-        shotnum, sni_dict, index_dict = \
-            do_shotnum_intersection(shotnum, sni_dict, index_dict)
-        self.assertTrue(np.array_equal(shotnum, [6, 7, 8]))
-        self.assertTrue(np.array_equal(sni_dict['Waveform'],
-                                       [True] * 3))
-        self.assertTrue(np.array_equal(index_dict['Waveform'],
-                                       [5, 6, 7]))
-
-    def test_two_controls(self):
-        """Test intersection behavior with two control devices"""
-        # test a case that results in a null result
-        shotnum = np.arange(1, 21, 1)
-        sni_dict = {
-            'Waveform': np.zeros(shotnum.shape, dtype=bool),
-            '6K Compumotor': np.zeros(shotnum.shape, dtype=bool)
-        }
-        index_dict = {
-            'Waveform': np.array([]),
-            '6K Compumotor': np.array([5, 6, 7])
-        }
-        sni_dict['6K Compumotor'][[6, 7, 8]] = True
-        self.assertRaises(ValueError,
-                          do_shotnum_intersection,
-                          shotnum, sni_dict, index_dict)
-
-        # test a working case
-        shotnum = np.arange(1, 21, 1)
-        shotnum_dict = {
-            'Waveform': shotnum,
-            '6K Compumotor': shotnum
-        }
-        sni_dict = {
-            'Waveform': np.zeros(shotnum.shape, dtype=bool),
-            '6K Compumotor': np.zeros(shotnum.shape, dtype=bool)
-        }
-        index_dict = {
-            'Waveform': np.array([5, 6]),
-            '6K Compumotor': np.array([5, 6, 7])
-        }
-        sni_dict['Waveform'][[5, 6]] = True
-        sni_dict['6K Compumotor'][[5, 6, 7]] = True
-        shotnum, sni_dict, index_dict = \
-            do_shotnum_intersection(shotnum, sni_dict, index_dict)
-        self.assertTrue(np.array_equal(shotnum, [6, 7]))
-        for key in shotnum_dict:
-            self.assertTrue(np.array_equal(sni_dict[key], [True] * 2))
-            self.assertTrue(np.array_equal(index_dict[key], [5, 6]))
+from ..hdfreadcontrol import HDFReadControl
 
 
 class TestHDFReadControl(TestBase):
